Pigeon IMU working, odometry baseline functional on 2020 drivetrain
We need to do a bit more calibration for odometry to be _*done*_, but it can handle driving around large sections of the field without drifting too badly. The Pigeon is absolutely amazing in how little it drifts, and judging by the amount of sensors built into it we have only begun utilizing its abilities. Also, now that a real robot is here, using ShuffleboardWidgets is now the right thing to do. It can run even with most of the subsystems missing, (which is a feature that maybe we should have in more of our code), and it has all the information we need (and we can add as much as we need). I also changed DrivetrainInterface to include a method to get yaw/pitch/roll (again, WIP).

It's not yet absolutely perfect (Electromechanical needed the robot back, so we only had 30 minutes to debug the odometry) but I'm confident that with more testing we can get it pretty accurate.

In terms of actual testing, we drove it in a super-rough square approximately 6 feet in width, and it registered reasonably accurate values at each corner. Of course, doing it with more precision would help, as would doing it in both directions, but there wasn't time. When we have long-term access to the robot, we can work on that.

Co-Authored-By: Chris-Chan <christianw1073@users.noreply.github.com>