Don't Go Breaking My Code
This code (for the prox sensors on the magazine) had previously been working, and was a part of a pull request called "Magazine Sensors Work". Most of the code from that push was deleted on this branch, and will delete all of the code from that original pull request on master if it is merged. Please do not change code that works.