Functional Autonomous Routine with Automatic Shooter Ranging and Indexing
The robot uses odometry and several InterpolationTables to move backward while indexing both the hood and the turret, and then automatically fire a power cell into the outer port.  This has actually tested. We need to do some tuning for a lot of the mechanisms, but this is like actually amazing.